+++ src/api/java/com/codetaylor/mc/artisanworktables/api/tool/reference/EnumWorktableToolType.java
@@ -56,7 +56,8 @@ public enum EnumWorktableToolType {
 
   // Universal
   UNIVERSAL_MORTAR("artisans_mortar", "Mortar"),
-  UNIVERSAL_KNIFE("artisans_knife", "Knife");
+  UNIVERSAL_KNIFE("artisans_knife", "Knife"),
+  UNIVERSAL_HATCHET("artisans_hatchet", "Hatchet");
 
   private final String name;
   private final String oreDictSuffix;